[Feature #134673091] Delete the old UI
