+++ codango/static/js/components/footer.jsx
