+++ codango/static/js/components/home.jsx
@@ -1,31 +1,33 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router';
-import { Col, Grid, Row } from 'react-bootstrap';
-import FormTabs from './formtabs.jsx';
+import React, {Component} from 'react';
+import Slider from './slider.jsx';
+import { StickyContainer, Sticky } from 'react-sticky';
+import SubMenu from './SubMenu.jsx';
+import Features from './features.jsx';
+import Feeds from './Feeds.jsx';
 
-class Home extends Component {
+export default class Home extends Component {
+constructor(props){
+    super(props);
+    this.state = {
+      showLogin: false
+    }
+  }
+  onStickyStateChange(isSticky) {
+    this.setState({ showLogin: isSticky})
+  }
   render() {
     return (
-      <Grid className="top-margin">
-        <Row className="show-grid">
-          <Col md={8}>
-            <div className="jumbotron">
-              <h1 id="index-h1">Join Our Community</h1>
-              <h3 className="section-text">Codango is a social networking
-                site that connects all types of developers allowing for
-                sharing resources, joining various communities and pair
-                programming
-              </h3>
-              <Link className="btn btn-primary btn-lg" to="/about">Learn more</Link>
-            </div>
-          </Col>
-          <Col md={4}>
-            <FormTabs />
-          </Col>
-        </Row>
-      </Grid>
+      <div style={{"background": '#fff'}}>
+        <Slider/>
+          <StickyContainer>
+            <Sticky onStickyStateChange={this.onStickyStateChange.bind(this)}>
+              <header>
+                <SubMenu showLogin={this.state.showLogin}/>
+              </header>
+            </Sticky>
+            <Feeds />
+          </StickyContainer>
+      </div>
     );
   }
 }
-
-module.exports = Home;
