--- codango/static/js/components/main-new.jsx
@@ -1,17 +1,0 @@
-import {render} from 'react-dom';
-import React, {Component} from 'react';
-import MenuNew from './menu-new.jsx';
-
-export default class MainNew extends Component {
-  constructor() {
-    super();
-  }
-
-  render() {
-    return (
-      <div>
-        <MenuNew/> {this.props.children}
-      </div>
-    )
-  }
-}
