+++ codango/static/js/components/main.jsx
@@ -1,20 +1,17 @@
-import { render } from 'react-dom'
+import {render} from 'react-dom';
 import React, {Component} from 'react';
-import Menu from './menu.jsx'
+import Menu from './menu.jsx';
 
 export default class Main extends Component {
-    constructor() {
-        super();
-    }
+  constructor() {
+    super();
+  }
 
-    render() {
-        return(
-            <div>
-                <Menu />
-                {this.props.children}
-            </div>
-        )
-    }
+  render() {
+    return (
+      <div>
+        <Menu/> {this.props.children}
+      </div>
+    )
+  }
 }
-
-module.exports = Main;
