+++ codango/static/js/components/team.jsx
@@ -1,199 +1,39 @@
-import {Grid} from 'react-bootstrap';
-var React = require("react");
-var Team = React.createClass({
+import React, { Component } from 'react';
+import { Grid, Row, Tabs, Tab } from 'react-bootstrap';
+import TeamMember from './team-member.jsx';
+import { teamMembers,  hallOfFame } from './team-directory.jsx';
+import { settings } from './react-slick-settings.jsx';
+const Slider = require('react-slick');
+
+class Team extends Component {
   render() {
     return (
-      <Grid className="top-margin">
-        <div className="heading">
-          <h1 className="page-title">Our Awesome Team</h1>
-        </div>
-        <div className="page-content">
-          <div className="row">
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars1.githubusercontent.com/u/3100850?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-cnnadi">
-                    <h4>Chidiebere Nnadi</h4>
-                  </a>
-                  <span>Team Lead</span>
-                </div>
-              </div>
-            </div>
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars2.githubusercontent.com/u/15629602?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/NdagiStanley">
-                    <h4>Stanley Ndagi</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars2.githubusercontent.com/u/12407721?v=3&s=400" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-mochieng">
-                    <h4>Margaret Ochieng</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="row">
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars3.githubusercontent.com/u/17270426?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-aabdulwahab">
-                    <h4>Abdulmalik Abdulwahab</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars2.githubusercontent.com/u/17288133?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-akiura">
-                    <h4>Alex Kiura</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars0.githubusercontent.com/u/18328313?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-snwuguru">
-                    <h4>Nwuguru Sunday</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="row">
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars1.githubusercontent.com/u/18309948?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-hoyeboade">
-                    <h4>Hassan Oyeboade</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="heading">
-            <h1 className="page-title">Hall Of Fame</h1>
-          </div>
-          <div className="row">
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars2.githubusercontent.com/u/13223950?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-ijubril">
-                    <h4>Issa Jubril</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars2.githubusercontent.com/u/13224175?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-ooshodi">
-                    <h4>Olufunmilade Oshodi</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars1.githubusercontent.com/u/9017229?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/Achile">
-                    <h4>Achile Egbunu</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-          </div>
-          <div className="row">
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars2.githubusercontent.com/u/15088852?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-ashuaib">
-                    <h4>Abiodun Shuaib</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars3.githubusercontent.com/u/13269579?v=3&s=460" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/andela-jngatia">
-                    <h4>Joan Ngatia</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-            <div className="col-sm-4">
-              <div className="team-members">
-                <div className="team-avatar">
-                  <img src="https://avatars2.githubusercontent.com/u/13224913?v=3&s=400" className="img-responsive"/>
-                </div>
-                <div className="team-desc">
-                  <a href="https://github.com/IniOluwa">
-                    <h4>Ini-Oluwa C. Fageyinbo</h4>
-                  </a>
-                  <span>Developer</span>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
+      <Grid id="team">
+        <Row className="show-grid">
+          <Tabs defaultActiveKey={1} id="team-tab">
+            <Tab eventKey={1} title="Team Members">
+              <Slider {...settings}>
+                {teamMembers.map((member, i) => {
+                  return (
+                    <div key={i}><TeamMember {...member}/></div>
+                  )
+                })}
+              </Slider>
+            </Tab>
+            <Tab eventKey={2} title="Hall Of Fame">
+              <Slider {...settings}>
+                {hallOfFame.map((member, i) => {
+                  return (
+                    <div key={i}><TeamMember {...member}/></div>
+                  )
+                })}
+              </Slider>
+            </Tab>
+          </Tabs>
+        </Row>
       </Grid>
-
-    );
+    )
   }
-});
-module.exports = Team;
+}
+
+export default Team;
